chasm/shb7:ibuff testing

+++ src/CHASM/chKEY.c
@@ -1,26 +1,46 @@
-/*/*//*//*//*//*//*//*//*//*//*//*//*//*//*/*/
-/*    chKEY                                 *
-*                                           *
-*     -sdl keyboard stuff                   *
-*                                           */
-// *   ---     ---     ---     ---     ---  *
-/*    LIBRE SOFTWARE                        *
-*                                           *
-*     Licenced under GNU GPL3               *
-*     be a bro and inherit                  *
-*                                           */
-// *   ---     ---     ---     ---     ---  *
-/* CONTRIBUTORS                             *
-*     lyeb,                                 *
-*                                           */
-/*/*//*//*//*//*//*//*//*//*//*//*//*//*//*/*/
+//**************************
+// CHKEY                   *
+// sdl keyboard stuff      *
+//                         *
+// LIBRE SOFTWARE          *
+// Licenced under GNU GPL3 *
+// be a bro and inherit    *
+//                         *
+// CONTRIBUTORS            *
+// lyeb,                   *
+//**************************
+
+
 
 #include "chKEY.h"
 
 //   ---     ---     ---     ---     ---
+// #:0x0;>
 
-void STWINKEY(uchar* k, uchar input)        {(*k)|= (1 << input);                           };
-void CLWINKEY(uchar* k, uchar input)        {(*k)&=~(1 << input);                           };
-int  GTWINKEY(uchar* k, uchar value)        { return ((*k)&value)==value;                   };
+void STWINKEY(
+  uint* k,
+  uint input) {
+
+  (*k)|=(1L<<input);
+
+};
+
+void CLWINKEY(
+  uint* k,
+  uint input) {
+
+  (*k)&=~(1L<<input);
+
+};
+
+int GTWINKEY(
+  uint* k,
+  uint value) {
+
+  return((*k)&value)==value;
+
+};
 
 //   ---     ---     ---     ---     ---
+// #:0x1;>
+
